# api/openapi/generated/thinkpixelmp.bundle.yaml
openapi: 3.1.0
info:
  title: ThinkPixelMP API
  version: 0.1.0-phase0
  description: Marketplace and agentic software supply-chain control-plane API.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: /v1
security:
  - bearerAuth: []
...
paths:
...
  /publishers:
    get:
      tags:
        - Publishers
      operationId: listPublishers
      summary: List tenant publishers
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          $ref: '#/components/responses/PublisherPage'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
    post:
      tags:
        - Publishers
      operationId: createPublisher
      summary: Create a tenant-local publisher
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/PublisherCreate'
      responses:
        '201':
          $ref: '#/components/responses/Publisher'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /publishers/{publisher_id}:
    get:
      tags:
        - Publishers
      operationId: getPublisher
      summary: Get a publisher
      parameters:
        - $ref: '#/components/parameters/PublisherId'
      responses:
        '200':
          $ref: '#/components/responses/Publisher'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
...
  /namespaces:
    get:
      tags:
        - Namespaces
      operationId: listNamespaces
      summary: List tenant namespaces
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          $ref: '#/components/responses/NamespacePage'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
    post:
      tags:
        - Namespaces
      operationId: createNamespace
      summary: Create a tenant namespace
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/NamespaceCreate'
      responses:
        '201':
          $ref: '#/components/responses/Namespace'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /namespaces/{namespace_id}:
    get:
      tags:
        - Namespaces
      operationId: getNamespace
      summary: Get a namespace
      parameters:
        - $ref: '#/components/parameters/NamespaceId'
      responses:
        '200':
          $ref: '#/components/responses/Namespace'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        maxLength: 4096
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 255
...
    PublisherId:
      name: publisher_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/uuidV7'
    NamespaceId:
      name: namespace_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/uuidV7'
...
  requestBodies:
    PublisherCreate:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/publisherCreate'
    NamespaceCreate:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/namespaceCreate'
...
  responses:
    PublisherPage:
      description: Page of publishers.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/publisherPage'
    NamespacePage:
      description: Page of namespaces.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/namespacePage'
...
    Publisher:
      description: Publisher representation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/publisher'
    Namespace:
      description: Namespace representation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/namespace-2'
...
    Problem:
      description: RFC 7807 Problem Details.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Problem:
      type: object
      additionalProperties: true
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          maxLength: 256
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
          maxLength: 2048
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          maxLength: 128
        request_id:
          $ref: '#/components/schemas/uuidV7'
    uuidV7:
      type: string
      format: uuid
      pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
    dnsToken:
      type: string
      minLength: 1
      maxLength: 63
      pattern: ^[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$
    publisher:
      type: object
      additionalProperties: false
      required:
        - id
        - tenant_id
        - slug
        - state
        - created_at
      properties:
        slug:
          $ref: '#/components/schemas/dnsToken'
        display_name:
          type: string
          minLength: 1
          maxLength: 256
        description:
          type: string
          maxLength: 4096
        id:
          $ref: '#/components/schemas/uuidV7'
        tenant_id:
          $ref: '#/components/schemas/uuidV7'
        state:
          type: string
          enum:
            - claimed
            - verified
            - suspended
            - revoked
        created_at:
          type: string
          format: date-time
    basePage:
      type: object
      additionalProperties: false
      required:
        - items
      properties:
        items:
          type: array
        next_cursor:
          type: string
          minLength: 1
          maxLength: 4096
    publisherPage:
      $ref: '#/components/schemas/basePage'
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/publisher'
    publisherCreate:
      type: object
      additionalProperties: false
      required:
        - slug
      properties:
        slug:
          $ref: '#/components/schemas/dnsToken'
        display_name:
          type: string
          minLength: 1
          maxLength: 256
        description:
          type: string
          maxLength: 4096
...
    namespace:
      type: string
      minLength: 1
      maxLength: 255
      pattern: ^[a-z0-9](?:[a-z0-9-]*[a-z0-9])?(?:/[a-z0-9](?:[a-z0-9-]*[a-z0-9])?)*$
    namespace-2:
      type: object
      additionalProperties: false
      required:
        - id
        - tenant_id
        - path
        - owner_publisher_id
        - created_at
      properties:
        id:
          $ref: '#/components/schemas/uuidV7'
        tenant_id:
          $ref: '#/components/schemas/uuidV7'
        path:
          $ref: '#/components/schemas/namespace'
        owner_publisher_id:
          $ref: '#/components/schemas/uuidV7'
        created_at:
          type: string
          format: date-time
    namespacePage:
      $ref: '#/components/schemas/basePage'
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/namespace-2'
    namespaceCreate:
      type: object
      additionalProperties: false
      required:
        - path
        - owner_publisher_id
      properties:
        path:
          $ref: '#/components/schemas/namespace'
        owner_publisher_id:
          $ref: '#/components/schemas/uuidV7'
...